Fix compiler error in updated VScode and spawing lots of children bug

See merge request bachelor/2020/group-107/bachelor!116

+++ bachelor/sim/src/entity/components/components.h
@@ -293,7 +293,7 @@ struct Reproduction
     float end_fertility{};
 
     /** General Pregnancy stat */
-    int mean_children_pp{1};
+    uint16_t mean_children_pp{1};
     float children_deviation{};
 
     /** Static fertility */
